Extract component and code clean up
