+++ ddp-workspace/package.json
@@ -35,7 +35,7 @@
     "core-js": "2.5.4",
     "hammerjs": "2.0.8",
     "kaop-ts": "3.1.2",
-    "ng-recaptcha": "^5.0.0",
+    "ng-recaptcha": "5.0.0",
     "ngx-cookie": "4.1.2",
     "npm-sass": "2.2.1",
     "package": "1.0.1",
