+++ ddp-workspace/projects/ddp-sdk/src/public-api.ts
@@ -63,6 +63,7 @@ export * from './lib/services/renewSessionNotifier.service';
 export * from './lib/services/scriptLoader.service';
 export * from './lib/services/serviceAgents/governedParticipantsServiceAgent.service';
 export * from './lib/services/languageService.service';
+export * from './lib/services/serviceAgents/invitationServiceAgent.service';
 
 export * from './lib/components/login/auth0-code-callback.component';
 export * from './lib/components/address/addressEmbedded.component';
