+++ ddp-workspace/projects/ddp-testboston/src/app/app-routes.ts
@@ -1,11 +1,8 @@
-import { UserRegistrationPrequalComponent } from './user-registration-prequal/userRegistrationPrequal.component';
-
 export const AppRoutes = {
     Activity: 'activity',
     ActivityId: 'activity/:id',
     ActivityLinkId: 'activity-link/:id',
     Error: 'error',
-    Prequalifier: 'prequalifier',
     LoginLanding: 'login-landing',
     Consent: 'consent',
     CovidSurvey: 'covid-survey',
