+++ ddp-workspace/projects/ddp-testboston/src/app/app-routing.module.ts
@@ -5,6 +5,7 @@ import { AppRoutes } from './app-routes';
 import { AppGuids } from './app-guids';
 
 import { WelcomeComponent } from './components/welcome/welcome.component';
+import { UserRegistrationPrequalComponent } from './components/user-registration-prequal/user-registration-prequal.component';
 
 import {
   AuthGuard,
@@ -12,7 +13,6 @@ import {
 } from 'ddp-sdk';
 
 import {
-  WorkflowStartActivityRedesignedComponent,
   LoginLandingRedesignedComponent,
   ActivityPageRedesignedComponent,
   DashboardRedesignedComponent,
@@ -21,16 +21,8 @@ import {
   PasswordRedesignedComponent,
   SessionExpiredRedesignedComponent
 } from 'toolkit';
-import { UserRegistrationPrequalComponent } from './user-registration-prequal/userRegistrationPrequal.component';
 
 const routes: Routes = [
-  {
-    path: AppRoutes.Prequalifier,
-    component: WorkflowStartActivityRedesignedComponent,
-    canActivate: [
-      IrbGuard
-    ]
-  },
   {
     path: AppRoutes.Consent,
     component: ActivityPageRedesignedComponent,
