+++ ddp-workspace/projects/ddp-testboston/src/app/components/user-registration-prequal/user-registration-prequal.component.html
@@ -1,0 +1,16 @@
+<div *ngIf="errorMessage" class="ErrorMessage">
+    <span>{{ errorMessage }}</span>
+</div>
+<form class="registration-prequal-form" [formGroup]="formGroup" (ngSubmit)="onSubmit()">
+    <mat-form-field class="example-input-field" floatLabel="always">
+        <mat-label>Enter invitation code</mat-label>
+        <input matInput uppercase formControlName="invitationId" pattern="[a-zA-Z0-9]{12}" maxLength="12"
+            placeholder="Your invitation code">
+    </mat-form-field>
+    <mat-form-field class="example-input-field" floatLabel="always">
+        <mat-label>What is your current mailing zip code?</mat-label>
+        <input matInput formControlName="zip" minlength="5" maxlength="5" placeholder="5 digit zip" pattern="[0-9]{5}">
+    </mat-form-field>
+    <re-captcha [siteKey]="config.recaptchaSiteKey" size="normal" formControlName="recaptchaToken"></re-captcha>
+    <button type="submit" [disabled]="formGroup.invalid" mat-raised-button color="primary">Submit</button>
+</form>
