+++ ddp-workspace/projects/ddp-testboston/src/app/components/user-registration-prequal/user-registration-prequal.component.scss
@@ -1,0 +1,4 @@
+.registration-prequal-form {
+    display: flex;
+    flex-direction: column;
+}
