+++ ddp-workspace/projects/ddp-testboston/src/app/components/user-registration-prequal/user-registration-prequal.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { UserRegistrationPrequalComponent } from './userRegistrationPrequal.component';
+import { UserRegistrationPrequalComponent } from './user-registration-prequal.component';
 
 describe('UserRegistrationPrequalComponent', () => {
   let component: UserRegistrationPrequalComponent;
