+++ ddp-workspace/projects/ddp-testboston/src/app/models/invitationCheckPayload.ts
@@ -2,5 +2,7 @@ export interface InvitationCheckPayload {
     recaptchaToken: string;
     invitationId: string;
     auth0ClientId: string;
-    qualificationDetails: {zipCode: string};
+    qualificationDetails: {
+        zipCode: string
+    };
 }
